Adds New Medibot Sprite and Faces.
Faces are overlays that can be applied.
Code is needed for such. Medibots have simply had their sprites replaced. I will need to change the other sprites sure as construction and some skins for them (etc.)

Credit to Goonstation for some of the faces ported from the Robobuddy.
As well as the idea of mashing both of em together.